fix(consensus): widen cumulative lottery-pool balance to u128
The lottery carryover pool drains at most one block reward per block
(the anti-grinding payout cap) but inflow per block is
emission_share + fee_pool, so sustained high-fee blocks grow the
cumulative carryover without bound. As a u64 with saturating_add it
would eventually saturate at u64::MAX, silently dropping value from
supply conservation.

Widen the cumulative carryover to u128 (LotteryPoolAccounting.available,
carryover_after, validate_block_lottery/apply_lottery stored_pool,
get_lottery_pool, and the META_LOTTERY_POOL on-disk encoding: 8->16-byte
LE). Per-block amounts (fee_pool, fee_burn, emission_share, payout,
payout_cap) stay u64 since each is bounded by a single block's fees or
reward. The anti-grinding cap is preserved: payout = min(available,
reward) computed in u128 and cast back to u64 losslessly.

The on-disk metadata format changes 8->16 bytes; it rides the planned
testnet reset (#323), so no migration path is needed. Also drops the now
redundant `lottery_pool as u128` cast in fuzz_add_block.rs (the getter
returns u128) and the matching cast in the ledger conservation test.

Closes #343

+++ botho/src/consensus/block_builder.rs
@@ -223,7 +223,7 @@ impl BlockBuilder {
     pub fn apply_lottery<F>(
         mut block: Block,
         candidates: &[LotteryCandidate],
-        stored_pool: u64,
+        stored_pool: u128,
         utxo_lookup: F,
         lottery_config: &LotteryFeeConfig,
     ) -> Block
